# lib/api-zod/src/generated/types/createPerformanceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DS Engineosys API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePerformanceRequest {
  employeeId: string;
  period: string;
  score: number;
  tasksCompleted: number;
  tasksFailed: number;
  efficiency: number;
  notes?: string | null;
}