# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DS Engineosys API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeStatus } from "./employeeStatus";

export interface Employee {
  id: string;
  name: string;
  email: string;
  employeeId: string;
  departmentId: string;
  departmentName: string;
  designation: string;
  joiningDate: Date;
  status: EmployeeStatus;
  performanceScore?: number | null;
  avatarUrl?: string | null;
}