# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DS Engineosys API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProductMarketStatus } from "./productMarketStatus";

export interface Product {
  id: string;
  name: string;
  category: string;
  sku: string;
  price: number;
  cost: number;
  stock: number;
  soldUnits: number;
  revenue: number;
  offerPercentage?: number | null;
  marketStatus: ProductMarketStatus;
  imageUrl?: string | null;
  createdAt: Date;
}