# lib/api-zod/src/generated/types/updateEmployeeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DS Engineosys API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEmployeeRequestStatus } from "./updateEmployeeRequestStatus";

export interface UpdateEmployeeRequest {
  name?: string;
  email?: string;
  departmentId?: string;
  designation?: string;
  status?: UpdateEmployeeRequestStatus;
}