# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DS Engineosys API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      summary: Get all employees
      parameters:
        - name: departmentId
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Register a new employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmployeeRequest"
      responses:
        "201":
          description: Employee created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get employee by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateEmployee
      tags: [employees]
      summary: Update employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmployeeRequest"
      responses:
        "200":
          description: Employee updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get all tasks
      parameters:
        - name: employeeId
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, in_progress, completed, failed]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /performance:
    get:
      operationId: getPerformanceRecords
      tags: [performance]
      summary: Get performance records
      parameters:
        - name: employeeId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Performance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerformanceRecord"
    post:
      operationId: createPerformanceRecord
      tags: [performance]
      summary: Create a performance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePerformanceRequest"
      responses:
        "201":
          description: Performance record created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PerformanceRecord"
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Add a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products/{id}/offer:
    post:
      operationId: applyProductOffer
      tags: [products]
      summary: Apply an offer to a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyOfferRequest"
      responses:
        "200":
          description: Offer applied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Employee:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        employeeId:
          type: string
        departmentId:
          type: string
        departmentName:
          type: string
        designation:
          type: string
        joiningDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive, on_leave]
        performanceScore:
          type: number
          nullable: true
        avatarUrl:
          type: string
          nullable: true
      required: [id, name, email, employeeId, departmentId, departmentName, designation, joiningDate, status]
    CreateEmployeeRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        employeeId:
          type: string
        departmentId:
          type: string
        designation:
          type: string
        joiningDate:
          type: string
          format: date
        status:
          type: string
          enum: [active, inactive, on_leave]
      required: [name, email, employeeId, departmentId, designation, joiningDate, status]
    UpdateEmployeeRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        departmentId:
          type: string
        designation:
          type: string
        status:
          type: string
          enum: [active, inactive, on_leave]
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        employeeId:
          type: string
        employeeName:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        dueDate:
          type: string
          format: date
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, title, employeeId, employeeName, status, priority, createdAt]
    CreateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        employeeId:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        dueDate:
          type: string
          format: date
          nullable: true
      required: [title, employeeId, status, priority]
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        priority:
          type: string
          enum: [low, medium, high, critical]
        dueDate:
          type: string
          format: date
          nullable: true
    PerformanceRecord:
      type: object
      properties:
        id:
          type: string
        employeeId:
          type: string
        employeeName:
          type: string
        period:
          type: string
        score:
          type: number
        tasksCompleted:
          type: integer
        tasksFailed:
          type: integer
        efficiency:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, employeeId, employeeName, period, score, tasksCompleted, tasksFailed, efficiency, createdAt]
    CreatePerformanceRequest:
      type: object
      properties:
        employeeId:
          type: string
        period:
          type: string
        score:
          type: number
        tasksCompleted:
          type: integer
        tasksFailed:
          type: integer
        efficiency:
          type: number
        notes:
          type: string
          nullable: true
      required: [employeeId, period, score, tasksCompleted, tasksFailed, efficiency]
    Product:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        sku:
          type: string
        price:
          type: number
        cost:
          type: number
        stock:
          type: integer
        soldUnits:
          type: integer
        revenue:
          type: number
        offerPercentage:
          type: number
          nullable: true
        marketStatus:
          type: string
          enum: [high_demand, moderate, low_demand, critical]
        imageUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, category, sku, price, cost, stock, soldUnits, revenue, marketStatus, createdAt]
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        sku:
          type: string
        price:
          type: number
        cost:
          type: number
        stock:
          type: integer
        soldUnits:
          type: integer
        marketStatus:
          type: string
          enum: [high_demand, moderate, low_demand, critical]
      required: [name, category, sku, price, cost, stock, soldUnits, marketStatus]
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        price:
          type: number
        stock:
          type: integer
        soldUnits:
          type: integer
        marketStatus:
          type: string
          enum: [high_demand, moderate, low_demand, critical]
    ApplyOfferRequest:
      type: object
      properties:
        offerPercentage:
          type: number
        reason:
          type: string
      required: [offerPercentage, reason]
...